fix(schema): convert autotrader_state.lot_units column to numeric and update OANDA_MAX_UNITS for Gold
- Alter lot_units column in Supabase from integer to numeric to support decimal lot sizes (0.01, 0.1, 0.5, 1.0)
- Update assert-schema.ts required type for lot_units to numeric
- Update OANDA_MAX_UNITS for XAU/USD in sizing.ts to 10,000 units (100 lots) instead of stale 100 units cap

+++ packages/core/src/assert-schema.ts
@@ -51,7 +51,7 @@ export const REQUIRED_COLUMNS: Record<string, string> = {
   'autotrader_state.id':                   'text PRIMARY KEY',
   'autotrader_state.mode':                 'user-defined enum',
   'autotrader_state.selected_instruments': 'text[]',
-  'autotrader_state.lot_units':            'integer',
+  'autotrader_state.lot_units':            'numeric',
   'autotrader_state.risk_profile':         'jsonb',
   'autotrader_state.updated_at':           'timestamptz',
   'autotrader_state.updated_by':           'text',
